chore: update openzeppelin version for `./packages/contracts`
Hardhat 3 will resolve the dependencies based on the current package.
Solidity files in Horizon where failing to compile because the
resolution mechanism jumped to `./packages/contracts` then onto the far
older version of `@openzeppelin/contracts`.

For the moment, I am bumping that version which leads to a clean `npx
hardhat compile`.